+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 ## Unreleased
 
 - Add `--version` flag.
+- Add support for selectively scoped ignoring with the `#scope` directive.
 - Fix alias resolution for `optionalDependencies` and `peerDependencies`.
 
 ## 0.3.11 (2025-09-17)